requirements.txt
alembic==1.16.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
geopandas==1.1.1
greenlet==3.2.3
haversine==2.9.0
idna==3.10
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
optuna==4.4.0
osmnx==2.0.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.1
pillow==11.3.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
seaborn==0.13.2
shapely==2.1.1
simpy==4.1.1
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
